HBASE-21389 Revisit the procedure lock for sync replication
